+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.2.3
+* Logging adjustments.
+
+### 0.2.2
+* Copy non image files so they can be sent to multiple folders.
+
 ### 0.2.1
 * Preserve all timestamps